add KillExcel flag

+++ DaxDrill/AddIn.cs
@@ -18,22 +18,24 @@ namespace DG2NTT.DaxDrill
 {
     public class AddIn : IExcelAddIn
     {
-        private Excel.Application xlApp = (Excel.Application)ExcelDnaUtil.Application;
-
+        // set to true to force Excel to close
+        public const bool KillExcel = false;
         public void AutoClose()
         {
-            xlApp.WorkbookDeactivate -= XlApp_WorkbookDeactivate;
+            var excelApp = (Excel.Application)ExcelDnaUtil.Application;
+            excelApp.WorkbookDeactivate -= XlApp_WorkbookDeactivate;
         }
 
         public void AutoOpen()
         {
-            xlApp.WorkbookDeactivate += XlApp_WorkbookDeactivate;
+            var excelApp = (Excel.Application)ExcelDnaUtil.Application;
+            excelApp.WorkbookDeactivate += XlApp_WorkbookDeactivate;
         }
 
         // kill Excel process in case objects are not properly released
         private void XlApp_WorkbookDeactivate(Excel.Workbook Wb)
         {
-            if (Wb.Application.Workbooks.Count == 1)
+            if (KillExcel && Wb.Application.Workbooks.Count == 1)
             {
                 Process.GetCurrentProcess().Kill();
             }
@@ -86,6 +88,7 @@ namespace DG2NTT.DaxDrill
                 if (sheet != null) Marshal.ReleaseComObject(sheet);
                 if (rngOut != null) Marshal.ReleaseComObject(rngOut);
                 if (excelApp != null) Marshal.ReleaseComObject(excelApp);
+                if (rngCell != null) Marshal.ReleaseComObject(rngCell);
             }
         }
